backend/requirements.txt
Flask
Flask-SocketIO
eventlet
redis
PyYAML
